+++ tests/Fixtures/Form/GrossPriceType.php
@@ -33,12 +33,12 @@ class GrossPriceType extends AbstractType
 
         $builder
             ->add('amount', IntegerType::class)
-            ->add('taxRate', ChoiceType::class, [
+            ->add($options['tax_rate_field_name'], ChoiceType::class, [
                 'choices' => [
                     '7%' => 7,
                     '19%' => 19,
                 ],
-            ])
+            ] + $options['tax_rate_field_options'])
         ;
 
         if ($options['extra_field']) {
@@ -58,5 +58,7 @@ class GrossPriceType extends AbstractType
         $resolver->setDefault('extra_field', false);
         $resolver->setDefault('include_button', false);
         $resolver->setDefault('map_extra_field', false);
+        $resolver->setDefault('tax_rate_field_name', 'taxRate');
+        $resolver->setDefault('tax_rate_field_options', []);
     }
 }